move timelockNamespace to api
