+++ lock-api/src/main/java/com/palantir/lock/TimelockNamespace.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.atlasdb.timelock;
+package com.palantir.lock;
 
 import org.immutables.value.Value;
