+++ timelock-impl/src/main/java/com/palantir/atlasdb/timelock/auth/api/Authorizer.java
@@ -16,7 +16,7 @@
 
 package com.palantir.atlasdb.timelock.auth.api;
 
-import com.palantir.atlasdb.timelock.TimelockNamespace;
+import com.palantir.lock.TimelockNamespace;
 
 public interface Authorizer {
     boolean isAuthorized(Client client, TimelockNamespace namespace);
