+++ timelock-impl/src/main/java/com/palantir/atlasdb/timelock/auth/api/NamespaceMatcher.java
@@ -16,7 +16,7 @@
 
 package com.palantir.atlasdb.timelock.auth.api;
 
-import com.palantir.atlasdb.timelock.TimelockNamespace;
+import com.palantir.lock.TimelockNamespace;
 
 public interface NamespaceMatcher {
     NamespaceMatcher NEVER_MATCH = ignored -> false;
