+++ timelock-impl/src/main/java/com/palantir/atlasdb/timelock/auth/impl/PrivilegeBasedAuthRequirer.java
@@ -20,7 +20,7 @@ import java.util.List;
 import java.util.Map;
 
 import com.google.common.collect.ImmutableList;
-import com.palantir.atlasdb.timelock.TimelockNamespace;
+import com.palantir.lock.TimelockNamespace;
 import com.palantir.atlasdb.timelock.auth.api.Client;
 import com.palantir.atlasdb.timelock.auth.api.NamespaceMatcher;
